+++ core/src/main/java/com/hujiugame/qingfeng/scene/impl/ConfigDisplay.java
@@ -1,0 +1,4 @@
+package com.hujiugame.qingfeng.scene.impl;
+
+public class ConfigDisplay {
+}